Created NoteController function to display notes in index.html page
New note-controller.js file containing a NoteController function which takes a note list model as a parameter, adds a note, and creates a note list view which passes in the note list model. The note controller has a method that gets HTML from the note list view and inserts it into the app element.

This was done by spiking code in order to know direction and next step is to TDD.