Passing test to check innerHTML property mtches the list
Checked that the innerHTML property of the app element contains the correct HTML string.

For this test, I mocked the NoteView and the NoteController. The reason for mocking the controller even though this is what is being tested is because I didn't want to actually update the DOM, just test that it would work.